# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "version": "1.0.0",
    "description": "API documentation for my Express application"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "get": {
        "summary": "Render the registration page",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Render the registration form"
          }
        }
      },
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username for the new user"
                  },
                  "email": {
                    "type": "string",
                    "description": "The email address for the new user"
                  },
                  "password": {
                    "type": "string",
                    "description": "The password for the new user"
                  },
                  "role": {
                    "type": "string",
                    "description": "The role of the new user (e.g., student, teacher)"
                  },
                  "mobile": {
                    "type": "string",
                    "description": "The mobile number for the new user"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Registration successful, redirect to login"
          },
          "400": {
            "description": "Registration failed due to validation errors"
          }
        }
      }
    },
    "/auth/login": {
      "get": {
        "summary": "Render the login page",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Render the login form"
          }
        }
      },
      "post": {
        "summary": "Log in an existing user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the user"
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the user"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Login successful, redirect to user role dashboard"
          },
          "401": {
            "description": "Invalid username or password"
          }
        }
      }
    },
...
    "/student/grades": {
      "get": {
        "summary": "View student's grades",
        "tags": [
          "Students"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved student's grades",
            "content": {
              "text/html": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "assignmentGrades": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "assignmentId": {
                            "type": "string",
                            "description": "The ID of the assignment"
                          },
                          "title": {
                            "type": "string",
                            "description": "The title of the assignment"
                          },
                          "grade": {
                            "type": "string",
                            "description": "The grade achieved in the assignment"
                          }
                        }
                      }
                    },
                    "quizGrades": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "quizId": {
                            "type": "string",
                            "description": "The ID of the quiz"
                          },
                          "title": {
                            "type": "string",
                            "description": "The title of the quiz"
                          },
                          "grade": {
                            "type": "string",
                            "description": "The grade achieved in the quiz"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "A message indicating the status of the grades"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An error occurred while fetching grades"
          }
        }
      }
    },
...
  },
...
}